requirements.txt
python-dotenv
selenium
Sphinx